feat: add infinite scroll to VirtualizedMasonryGrid
- moved complex logic from VirtualizedMasonryGrid component to useGridLayout and useVisibleItems custom hooks
- added descriptions of interface APIs of VirtualizedMasonryGrid and MasonryGrid components
- moved GridItem and GridItemPosition type declarations to shared types folder
