+++ src/components/MasonryGrid.tsx
@@ -1,10 +1,21 @@
 import { type ReactNode, Children } from 'react';
 import { GridContainer, GridItem } from './MasonryGrid.styled.ts';
-import type { ColumnsConfig } from '../types/masonry.ts';
+import type { GridColumnsConfig } from '../types/masonry.ts';
 
 interface MasonryGridProps {
+  /**
+   * ReactNode to render for each item.
+   */
   children: ReactNode;
-  columns?: number | ColumnsConfig;
+
+  /**
+   * Number of columns in the grid or a GridColumnsConfig object.
+   */
+  columns?: number | GridColumnsConfig;
+
+  /**
+   * Gap between grid items in pixels.
+   */
   gap?: number;
 }
