+++ src/components/VirtualizedMasonryGrid.tsx
@@ -1,30 +1,44 @@
-import { type ReactNode, useEffect, useState, useRef, useCallback, forwardRef } from 'react';
+import { ReactNode, useEffect, useRef, forwardRef } from 'react';
 import { useTheme } from 'styled-components';
-import debounce from 'lodash.debounce';
-import { GridContainer, GridItem } from './VirtualizedMasonryGrid.styled.ts';
-import { getColumns } from '../utils/masonry.ts';
-import type { ColumnsConfig } from '../types/masonry.ts';
+import { GridContainer, GridItem as StyledGridItem } from './VirtualizedMasonryGrid.styled.ts';
+import { useGridLayout } from '../hooks/masontry/useGridLayout.ts';
+import { useVisibleItems } from '../hooks/masontry/useVisibleItems.ts';
+import type { GridColumnsConfig, GridItem } from '../types/masonry.ts';
 
-type BaseGridItem = {
-  id: string | number;
-  height: number;
-  width: number;
-};
-
-interface GridItemPosition {
-  translateX: number;
-  translateY: number;
-  width: number;
-  height: number;
-}
-
-interface VirtualizedMasonryGridProps<T extends BaseGridItem> {
+interface VirtualizedMasonryGridProps<T extends GridItem> {
+  /**
+   * Array of items to display in the grid.
+   */
   items: T[];
+
+  /**
+   * Function that returns a ReactNode to render for each item.
+   */
   children: (item: T, index: number) => ReactNode;
-  columns?: number | ColumnsConfig;
+
+  /**
+   * Number of columns in the grid or a GridColumnsConfig object.
+   */
+  columns?: number | GridColumnsConfig;
+
+  /**
+   * Gap between grid items in pixels.
+   */
   gap?: number;
+
+  /**
+   * Number of extra items to render.
+   */
   overscan?: number;
+
+  /**
+   * Function to load more items (infinite scrolling).
+   */
   loadMore?: (index: number) => void;
+
+  /**
+   * Indicating whether there are more items to load.
+   */
   hasMore?: boolean;
 }
 
@@ -38,7 +52,7 @@ const LoadingFallback = forwardRef<HTMLDivElement>((_, ref) => {
   );
 });
 
-const VirtualizedMasonryGrid = <T extends BaseGridItem>({
+const VirtualizedMasonryGrid = <T extends GridItem>({
   items,
   children,
   columns = DEFAULT_COLUMNS,
@@ -49,127 +63,25 @@ const VirtualizedMasonryGrid = <T extends BaseGridItem>({
 }: VirtualizedMasonryGridProps<T>) => {
   const containerRef = useRef<HTMLDivElement | null>(null);
   const loaderRef = useRef<HTMLDivElement | null>(null);
-
-  const [positions, setPositions] = useState<GridItemPosition[]>([]);
-  const [visibleItems, setVisibleItems] = useState<number[]>([]);
-  const [containerHeight, setContainerHeight] = useState(0);
-
   const theme = useTheme();
 
-  const updateLayout = useCallback(() => {
-    if (!containerRef.current) return;
+  const { positions, containerHeight } = useGridLayout({
+    items,
+    columns,
+    gap,
+    containerRef,
+    breakpoints: theme.viewport.breakpoints,
+  });
 
-    const columnsCount =
-      typeof columns === 'number' ? columns : getColumns(columns, theme.viewport.breakpoints);
+  const visibleItems = useVisibleItems({
+    positions,
+    containerRef,
+    overscan,
+    gap,
+    itemsLength: items.length,
+  });
 
-    const containerWidth = containerRef.current.offsetWidth;
-    const columnWidth = (containerWidth - gap * (columnsCount - 1)) / columnsCount;
-    const columnHeights = Array(columnsCount).fill(0);
-
-    const calculatedPositions = items.map((item) => {
-      const shortestColumn = columnHeights.indexOf(Math.min(...columnHeights));
-      const translateX = shortestColumn * (columnWidth + gap);
-      const translateY = columnHeights[shortestColumn];
-
-      const aspectRatio = item.height / item.width;
-      const itemHeight = columnWidth * aspectRatio;
-      const itemWidth = columnWidth;
-
-      columnHeights[shortestColumn] += itemHeight + gap;
-
-      return {
-        translateX,
-        translateY,
-        width: itemWidth,
-        height: itemHeight,
-      };
-    });
-
-    setPositions(calculatedPositions);
-    setContainerHeight(Math.max(...columnHeights));
-  }, [items, gap, columns, theme.viewport.breakpoints]);
-
-  const updateVisibleItems = useCallback(() => {
-    if (!positions.length || !containerRef.current) return;
-
-    const containerRect = containerRef.current.getBoundingClientRect();
-    const containerTop = containerRect.top + window.scrollY;
-
-    const scrollTop = window.scrollY;
-    const windowHeight = window.innerHeight;
-
-    const start = scrollTop;
-    const end = scrollTop + windowHeight;
-
-    // create an array with positions and indices
-    const positionsWithIndex = positions.map((pos, index) => ({
-      index,
-      translateY: pos.translateY,
-      height: pos.height,
-    }));
-
-    // sort the array by translateY
-    positionsWithIndex.sort((a, b) => a.translateY - b.translateY);
-
-    const viewportIndices: number[] = [];
-    for (const item of positionsWithIndex) {
-      const itemTop = containerTop + item.translateY;
-      const itemBottom = itemTop + item.height;
-
-      if (itemBottom + gap >= start && itemTop - gap <= end) {
-        viewportIndices.push(item.index);
-      } else if (itemTop - gap > end) {
-        // break if items are sorted
-        break;
-      }
-    }
-
-    if (viewportIndices.length === 0) {
-      setVisibleItems([]);
-      return;
-    }
-
-    const startIndex = Math.max(0, viewportIndices[0] - overscan);
-    const endIndex = Math.min(
-      items.length - 1,
-      viewportIndices[viewportIndices.length - 1] + overscan,
-    );
-
-    setVisibleItems(Array.from({ length: endIndex - startIndex + 1 }, (_, i) => startIndex + i));
-  }, [positions, overscan, items.length, gap]);
-
-  useEffect(() => {
-    updateLayout();
-
-    const debouncedResizeHandler = debounce(() => {
-      updateLayout();
-    }, 200);
-
-    window.addEventListener('resize', debouncedResizeHandler);
-    return () => {
-      debouncedResizeHandler.cancel();
-      window.removeEventListener('resize', debouncedResizeHandler);
-    };
-  }, [updateLayout]);
-
-  useEffect(() => {
-    updateVisibleItems();
-
-    const throttledScrollHandler = debounce(() => {
-      updateVisibleItems();
-    }, 50);
-
-    window.addEventListener('scroll', throttledScrollHandler);
-    return () => {
-      throttledScrollHandler.cancel();
-      window.removeEventListener('scroll', throttledScrollHandler);
-    };
-  }, [updateVisibleItems]);
-
-  useEffect(() => {
-    updateVisibleItems();
-  }, [positions, updateVisibleItems]);
-
+  // infinite scrolling to load more items
   useEffect(() => {
     if (!loadMore || !hasMore || containerHeight === 0) return;
 
@@ -198,7 +110,7 @@ const VirtualizedMasonryGrid = <T extends BaseGridItem>({
           const item = items[index];
 
           return (
-            <GridItem
+            <StyledGridItem
               role="list-item"
               key={`${item.id}-${index}`}
               $translateX={pos.translateX}
@@ -207,7 +119,7 @@ const VirtualizedMasonryGrid = <T extends BaseGridItem>({
               $height={pos.height}
             >
               {children(item, index)}
-            </GridItem>
+            </StyledGridItem>
           );
         })}
       </GridContainer>
