+++ src/hooks/query/usePhoto.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
-import { getPhoto } from '../api/pexels';
-import type { GetPhotoParams, GetPhotoResponse } from '../types/pexels';
+import { getPhoto } from '../../api/pexels.ts';
+import type { GetPhotoParams, GetPhotoResponse } from '../../types/pexels.ts';
 
 export const usePhoto = (params: GetPhotoParams, options?: { enabled?: boolean }) => {
   return useQuery<GetPhotoResponse, Error>({
